re-structure tableaux specific databasequeries into .database.model.tableaux
